framework: tlvf: test: add value-checking for channel scan results
This test is very different from the others. It creates a
ChannelScanResult TLV from an existing buffer, and checks that each
field has the expected value one by one.

It is different from the others in the sense that it only checks a
specific TLV.

This TLV was chosen because it contains conditional parameters, and is
a perfect case to test the new conditional parameters feature in TLVF.

The TLV includes a class with an optional parameter that is not
present, and one where it is present. This allows us to check that we
parse it correctly in both cases. It comes from a Channel Scan Result
frame from the R2 certification testbed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>