controller: agent_monitoring_task: fix null pointer dereference
For devices not implementing Profile2 (e.g. the devices from the R1
certification testbed), the Profile2 Radio Metrics will not be
present.

For these devices, we need to ignore the missing TLV.

The existing check only reported the missing TLV for R2-or-later
devices. To fix it, split the condition into two:

- first, check if the TLV is present.

- if it is NOT present, then check if the device is a R2-or-later
device.

- if it is R2-or-later, then log an error and return. If it is not,
silently return (as we expect R1-only devices to not have the TLV).

Closes PPM-1574

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>