controller: agent_monitoring_task: swapped unicast and broadcast metrics
The metrics for broadcast and unicast bytes were swapped when they
were added in 38ed74f026dd93a5eb5d22c6fa9eebdeba924d30, so fix it.

While at it, also change the field and assignment order of the
metrics to match the order they are in in the TLV. This is not
required, but it's easier to check that the assignments are correct
when they are in the same order.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>