tests: environment.py: sort imports
According to PEP8 (see [1]), the order should be:

1. Standard library imports.
2. Related third party imports.
3. Local application/library specific imports.

When adding a new import, it's difficult to know where to put it,
since the current order is not correct.

Sort the imports and add comments, so that other developers know where
to add new imports.

Note that linters (such as pylint) also report the wrong import order.

[1]: https://www.python.org/dev/peps/pep-0008/#imports

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>