agent: Fix unsafe Backhaul Manager members access from two threads
Before the instantiation of the Backhaul Manager thread Class, we are
creating the CMDU server and the Broker Client with the EventLoop
object of the Platform Manager.
On the Backhaul Manager, we assign the Backhaul FSM timers and
Tasks Timers with the local EventLoop Object.

This has a hazardous impact.
It means that the handlers of the messages are running in the Platform
Manager thread context whereas the FSM of the Backhaul and the Task
run in a different context.
So both threads have access to the same variables without mutex
protection.

Fix it by moving the CMDU server and the Broker Client instantiation into
the Backhaul Manager, giving them the local event loop instance.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>