bwl: dummy, nl80211: handle AP_DISABLED and AP_ENABLED in AP hal
The dummy and nl80211 didn't have handlers for AP_DISABLED and
AP_ENABLED yet.

The implementations are the same as for dwpal.

For the dummy bwl, additionally set radio_state directly, since we
cannot "read" the state in refresh_radio_info (because there is no
radio to read the state from).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>