bcl: Create thread init function
There are some actions done in the constructor/init function of some
thread modules in the code that prints stuff.
The problem with that is that all the print will be printed to the wrong
file in case the thread has a different log file.
Add a thread_init function, that will take place in the thread context.
Since the thread_init is preferred over the regular init function, make
it pure virtual.
The regular init function remains for compatibility with legacy code
that still uses it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>