bwl: nl80211: set radio_state
The state was previously not set when getting the radio information.

To make it easier, refactor the state map from dwpal and
place it in base_wlan_hal, so that we can use it for the other BWLs.

The dummy implementation will be treated separately: instead of
setting the state when reading the radio info, we will set it when
disabled/enabled is received (in a later commit).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>