bwl: RadioInfo: remove radio_enabled
There is some duplication between radio_state and
radio_enabled. Keeping the two makes it more likely that we forget to
update one (a good example is that radio_enabled was previously set
for the nl80211 bwl but not radio_state).

Remove radio_enabled as it's used in only one place, where we can
just as well use radio_state directly.

For the dummy BWL, radio_enabled was hard-coded to true in
refresh_radio_info(). Because in a later commit we will allow to
disable and enable the radio, initialize radio_state only once in
attach().

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>