agent: Move thread component creation to thread init
If the instantiation of some thread components like CMDU server or client
is done in the constructor or init function, the logs of it are being printed
on the context and log file of the module creator thread context and not
on the created thread context.
Move module component instantiation and initialization from the constructor
of the Monitor and the Backhaul manager threads to the tread_init() functions.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>