bcl: Name sockets
When opening a socket we would like to name that socket,
so if something happens that requires print we will have the
ability to print the socket name.

Add a 'name' member to all of the sockets related classes,
which will be used when registering the handlers of that socket.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>