bcl: Name timers
When adding a timer there is no way to know which timer is it on the
log file.
Add 'timer name' parameter to the add_timer() function, so whenever an
error event will occur on that timer file descriptor we would be able to
know which timer was it.

Name all timers in the code.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>