bcl: Allow setting a file descriptor EventLoop handlers name on runtime
At a client connection, it is not possible to know who that client is.
This function allows setting the file descriptor name in the EventLoop
handlers later on when the client identified so in case of EventLoop
error or a client disconnect event it will be possible to know on which
client the event has occurred.

Add the required functionality to allow assigning the client names in
runtime, and set them on the backhaul manager and the platform manager
which are the only places using event loop that accept incoming
connection.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>