agent: ignore renew if not yet connected
When we receive an autoconfig renew message, we immediately go to the
STATE_JOIN_MASTER state so that we start the autoconfiguration again.

However, if we happen to receive a renew before we're connected to the
controller[1], we're going to attempt the autoconfiguration before
master_socket is set.

When this happens, the following line is visible in the logs:
master_socket == nullptr

To fix it, ignore autoconfig renew messages if master_socket is not
set yet.

Note that the master socket will be removed as part of PPM-350, so
this should be changed again in the future to use
ap_autoconfiguration_completed instead.

Closes PPM-1581

[1] This can for example happen in certification tests where a
controller could still be configured from a previous test, right
before the dev_reset_default command is sent to every device. See the
following two test runs for example:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1533400210
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1543644092

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>