framework: tlvf: add allocation for conditional elements
The conditional elements are not variable length lists, but they rely
on the same idea: when allocating, move all the pointers for the rest
of the buffer by the length of the allocated field.

Add allocation methods similar to variable length lists.

A follow-up commit will call addClassVarLenMethods so that the
allocation methods are actually created.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>