framework: tlvf: add support for conditional elements
When an element is marked as conditional, we:

- use pointers to access it. If the element is not present, the getter
will return a nullptr.

- use setters instead of accessing them directly. Setting the element
will allocate room for it on the buffer if needed, set the element,
and set the presence field. This uses the same mechanism as for
variable length list, with an "alloc_" method.

- can set it multiple times, the element will not (and must not) be
allocated a second time. For that we keep a flag: m_<field>_allocated.

- only need to swap the element if the presence field is set.

To keep it simple, for now only arithmetic types can be
conditional. For the other cases, add a call to abort() to warn the
user that it's not supported.

When creating an outgoing TLV, it's also currently assumed that the
value for the presence field is 0 when the conditional field is not
present (which matches the only use-case we have for now: "load
element" of the Channel Scan Result TLV).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>