association_handling_task: fix rssi measurement flow
As part of an investigation of a timeout on receiving the
ACTION_CONTROL_CLIENT_RX_RSSI_MEASUREMENT_RESPONSE after sending the
request in the association_handling_task, the following issues in the
flow were found and fixed:
- the REQUEST_RSSI_MEASUREMENT state changes to the FINISH state even if
timeout is reached for the above-mentioned response. Removed the
change-state to FINISH from the REQUEST_RSSI_MEASUREMENT state handling
and move it to the handling of the response message (until now response
handling was empty).
- The above-mentioned change fixes the response timeout handling and
will result in retries as expected - since the timeout handler expects
the state to be REQUEST_RSSI_MEASUREMENT for the above-mentioned event
timeout.
- Added a call to finish() of the task when max-attempts of
rssi-measurements is reached.

Signed-off-by: Adam Dov <adov@maxlinear.com>