tools: commands: build.py: add support for --parallel
When building, we sometimes need to build with a lower amount of
parallelism than the number of cores (for example because the amount
of available RAM is low compared to the number of cores).

Add a --parallel option to the build script, that is passed along to
cmake.

Additionally, Ninja already uses the maximum number of cores by
default. So when the generator is Ninja and no --parallel option was
specified, don't set to the number of cores and let Ninja use its
defaults.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>