boardfarm: added function on capi.py to get tlv length
In order to avoid the constant need to input the tlv length for a
given value, a function was added to calculate it on capi.py,
Its worth mentioning that in some cases the values sent don't
correspond to their actual decimal conversion.
For example, the hex value 0x000000110022 which translates to a
1114146 decimal value, should only occupy 3 bytes. Instead, it should
be readen as 0x00 00 00 11 00 22 and have the length of 6 bytes
because any of the zeroed octets can, in some circunstances, contain
a different value.

Signed-off-by: David <davidmgafonso@gmail.com>