boardfarm: get tlvs from yaml files
Setting messages and tvls as an hardcoded hex
number makes it hard to follow up on code logic.
In order to keep the existing tlvs accessible
as a python object, the yaml files from the
framework folder are to be scanned and mapped
into a usable dictionary so its content can be
later used on the tests.

These files would normally be parsed as
file name > (type name we want to use).
The namespace/type we want to use, is in the
fact the file name as well (in most cases).
To avoid having repeated keys, which would
mean we would have to deeper on a dict to get
a usable value (example: self.ieeeTlvs['eTlvType']
['eTlvType']['value_we_want']) the first first
key is omitted if it corresponds to the file name.

Signed-off-by: David <davidmgafonso@gmail.com>