bcl: Add suffix argument to create_vlan_interface()
Add a 'suffix' as an option argument on create_vlan_interface().
This argument enables the creation of a VLAN with a fixed name, which is
not dependent on the VLAN ID.
This is helpful in situations when we would want to "update" a VLAN ID
on a VLAN interface. It is not possible to with 'ip link' command so the
way to do it is to remove the interface and recreate it. It is very
convenient if the interface has a constant name.

In Addition, some operation could be depended on the VLAN creation.
Running in detach mode, leads to a race between the operation and could
cause to an an undesired outcome. To prevent it, change it to detached mode
so all the operations will be done in a synchronic order.

While at it, increase characters reservation for the command, which was
too lower.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>