agent: Add VLAN interface linked to every bAP/bSTA interface
Since multicast messages are not bridged, and instead of, being sent to
all interfaces, they will lack a VLAN tag. To overcome it, add a VLAN
interface with the Primary VLAN on the backhaul interface.
Only a Primary VLAN is needed since it is the only VLAN that IEEE 1905.1
messages could be sent on.
Use ".pvid" suffix so it will be easy to change the VLAN ID if changed
by the Controller.
Add the new VLAN interface linked to every Profile2-bAP and bSTA
interfaces.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>