agent_config: add polling-rate and window-poll-count params
Preparative commit.
Currently the monitor polling rate is hard-coded, but for setups where
we expect many clients to be connected we might want to reduce it to
reduce the load on the hostapd and the CPU.

The above-mentioned parameters are optional and can be used to override
the default values used for polling station statistics by the monitor.

Default values:
polling-rate: 250 msec
window-poll-count: 4 times.

PPM-1525.

Signed-off-by: Adam Dov <adov@maxlinear.com>