message_bml: change the bml buffers to use uint8_t
A removal of a MACRO of GET_MESSAGE_POINTER used an API to access a
location in the buffer using obj->buffer(index) but that API was
implemented by char* APIs in the tlvf.
The reason for this is that the messages are wrongly defining the
buffers as char arrays instead of uint8_t.

Changed the BML messages that define a buffer to use it as uint8_t array
and not char array.

Signed-off-by: Adam Dov <adov@maxlinear.com>