transport: Fix transport interface configuration
The transport allows to add and remove interfaces from its interfaces
list. This interface list is used when the transport wishes to send
a packet to all interfaces.
Although the configuration mechanism exists, it is implemented poorly:
1. It is not possible to remove an interface.
2. It is not possible to configure an interface that is outside of
a bridge.

Add new members to the Metadata struct which is used for configuring
interfaces by the broker, to allow adding and removing interfaces which
are inside or outside a bridge.

Change the function 'update_network_interfaces` to receive separately,
list of interfaces to remove and to add/update.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>