agent: Configure VLAN interface on the transport
When the Transport forward locally generated packets or
"relayed multicast" packets coming from the network it does not forward
them to the bridge because the bridge will be not forward them due to
an ebtables rule that drops them.
To successfully forward the packet, the Transport instead send them to
all interfaces directly (bypassing the bridge).
When using traffic separation, bypassing the bridge is causing to
packets to not get tagged, since it is the bridge that is responsible to
add the taggings.

A previous commit added a new VLAN interface to solve this, but the
transport was not familiar with the new interface.
Add configuration message to the transport after adding or removing
the new VLAN interface.
Remove the original wireless backhaul interface from the transport, so
a packet will be not sent twice from both interfaces.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>