agent: backhaul_manager: check slave socket before using it
In some cases, we try to forward a cmdu even when there is no slave to
forward to (e.g. right after a dev_reset_default command). This leads
to a segmentation fault.

Fix it by checking if there is a slave socket and if it is valid.

Fixes PPM-1602

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>