agent: basic capabilities: read the number of BSSes from the Agent DB
When filling the radio basic capabilities (e.g. in the M1 message, in
AP metrics responses, ...) we used to hard-code the number of BSSes to
4.

This was a problem because in reality, on the devices we test there
are 4 BSSes available on Axepoint, and 2 on Turris Omnia.

Instead of hard-coding the value, count their number from the Agent
DB.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>