common: tlvf: cACTION_APMANAGER_JOINED_NOTIFICATION: add vap_list
This is a preparatory commit.

When the AP manager joins, if there are any VAPs already available we
need to include them in the notification, so that the son_slave_thread
knows what VAPs are available.

Add a vap list to the notification, and update the auto-generated
files.

In a later commit, the new list will be used by the AP manager.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>