agent: son_slave_thread: factor out updating the VAP information
This is a preparatory commit.

We currently update the VAP information when we receive
"ACTION_APMANAGER_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION" (i.e. when
autoconfiguration is complete).

In a later commit, we will also update them when the AP manager
joins. To make it easier, refactor it in its own method.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>