ci: altran testbed not using ninja and changed to 4 cores
Previously the altran testbed would get stuck while
build prplmesh (PPM-1606). After some investigation
it seems that the board would become unresponsive while
using all of its cores on a high load task.
To overcome this, an option was made of just using one
of the cores while building prplmesh. After some tries,
it seemed that even while working with just one core,
the RPI would sometimes crash for not being able to complete
a task in 120seconds. It was then suggested that disabling
ninja would ease the load and make the build viable. This
last suggestion seems to work so far.

Signed-off-by: David <davidmgafonso@gmail.com>