ci: altran testbed needs param changed
Previously as there was no "needs" param, the
run-tests-on-altran-glinet-testbed would wait for
every build to be done before running the tests.
Since the image for prplmesh is being built on the
testbed RPI and not relying on any other part of the
pipeline, we can just skip all these builds and start
running the tests once we have a build for the glinet
ready.

Signed-off-by: David <davidmgafonso@gmail.com>