bwl: nl80211::update_vap_credentials: require bssid only in bss section
bwl nl80211 failed to update vap credentials of main vap: missing the
bssid field in the mainBss interface section of radio's hostapd config file.

fix: bssid field is only defined (and would be checked) in the bss configuration
section. This latter is created only for second and following BSSes (multi-bss)

https://jira.prplfoundation.org/browse/PPM-1638

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>