ci: fix dco check when source branch has been merged to target before
When we try to merge a source branch that has been merged into the
target branch at an earlier point in the Git history (e.g. merging a
stable branch a second time into master), the dco-check consistently
fails.

This is because in this case, this command returns an empty value: git
merge-base --fork-point origin/$CI_MERGE_REQUEST_TARGET_BRANCH_NAME

See [1].

Eventually we need to turn the workarounds we use here into actual
fixes in the upstream dco-check project, instead of maintaining them
on our own. For now they only have been reported ([2]).

While at it, also remove the `-b` option, as it should be guessed
automatically.

This might fix the issue reported in PPM-932 (it's difficult to be
sure, as we don't have much information about the failure).

[1]: https://github.com/christophebedard/dco-check/issues/113
[2]: https://github.com/christophebedard/dco-check/issues/114

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>