controller: check for stats_info before de-referencing it
The stats_info pointer was accessed without being checked.

While it shouldn't happen in theory (because it's initialized by the
constructor depending on the node type and never reset to nullptr
afterwards), in practice it can happen if nodes are created with the
wrong type and/or their type gets changed.

Check the pointer, log an error and return if it's null. This won't
fix the problem of nodes that get their type changed when they
shouldn't, but it will at least prevent the controller from crashing
in this specific case.

This should fix at least one segmentation fault discovered while
investigating PPM-1635.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>