controller: db: initialize node members
Some members of `node` were initialized conditionally, depending on
their type.

This is a bad idea, as it leaves the door open for undefined
behaviors (which if we're lucky will trigger a segmentation fault).

In PPM-1635, the `stats_info` member was accessed while being
uninitialized (possibly because the node gets its type changed), which
caused a segmentation fault:

son::db::notify_disconnection (this=0xbea4a2ac, client_mac=...) at /home/rme/projects/prplwrt/build_dir/target-arm_cortex-a9+vfpv3-d16_musl_eabi/prplMesh-2.2.0/controller/src/beerocks/master/db/db.cpp:4190
4190        if (!m_ambiorix_datamodel->set(path_to_eventdata, "BytesSent", n->stats_info->tx_bytes)) {

A similar issue was already fixed in [1] by checking the `stats_info`
pointer before de-referencing it.

Instead of adding checks for `stats_info` in this place and
potentially others where it's also used, this time remove the
conditional initialization, and always initialize it.

A potential issue with this change is if an existing piece of code
relies on the initialization of the pointer in a conditional (as a
round-about way to determine the node type for example). From a quick
look though, such piece of code doesn't seem to exist.

While at it, do the same for `hostapd` and its own `stats_info`
member.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/2670

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>