agent: Move the backhaul and controller socket out of the radio context
The unified agent should have a single socket to the backhaul manager.
Move the backhaul manager and the controller(master_socket) sockets to
the global agent context instead of the radio context as it is now.
As part of this change:
1. Move the "is_backhaul_disconnected" flag as well since it is coupled
in usage with the sockets.
2. Remove "fronthaul_iface" argument from functions that used it to get
the socket.
3. Remove the RadioManager "get_controller_socket_iface()" as it is not
required anymore for getting the socket to the Controller.
4. Replace "fronthaul_iface" argument with an empty string on some
send_cmdu_to_controller() calls where it is not needed for filling
a specific radio mac on the beerocks VS message header.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>