agent: Remove is_backhaul_reconf flag
is_backhaul_reconf used to indicate whether there was a reconnection
event on the backhaul manager while we haven't finish the
auto-configuration phase.
This flag have effect on a flow in join message handler in the
Controller which using the obsolete channel selection task to maybe
switch the radio channel before activating the radio node on the
database.
Since this flow is no longer active, remove this flag, and set all the
database fields as if there wan't a reconf before activating the radio.

Also remove the flag from the message yaml file and its auto-generated
files.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>