agent: Use radio mac on beerocks message header
Some messages sent from the Backhaul manager, including from the tasks,
are destined for a specific radio to perform some operation.
Since from now on, we will have only a single socket between the
backhaul manager and the Agent, we need to explicitly mention which
radio should perform the operation.
Unlike the messages on the previous commit that we added an interface
name to the message, on these messages we use the radio_mac in the
beerocks header.
The reason for using the beerocks header instead of adding an interface
name to the message is because the messages in this commit should move
anyway as part of the tasks transition to the Agent, and we will send
the messages directly to the radio thread (ap_manager/monitor).

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>