agent: Remove/ Comment out useless code
Remove ACTION_BACKHAUL_CHANNEL_SCAN_DUMP_RESULTS_REQUEST message and its
handler on the Agent thread since no one sends this message.

Comment out the Handler of ACTION_BACKHAUL_DL_RSSI_REPORT_NOTIFICATION
message in the Agent thread. The handler is forwarding the message to
the Controller, but the handler in the Controller is commented out.
It is suggested on the Controller that we might need it and it is
temporarily commented out. So the same is done on the Agent side.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>