agent: Change how we work with ZWDFS radio
Previously the state WAIT_FOR_FRONTHAUL_THREADS_JOINED has been
performed in the radio context.
Since now it is performed in the global FSM context, we can't send the
ZWDFS detection message to the backhaul manager in that state.
The solve this, the detection message is being sent in the handler of
the AP_MANAGER_JOINED where we get this flag.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>