agent: Backhaul manager measurements responses
The messages handlers of the of backhaul measurement flows sends
responses to the Controller.
Since now we have a single socket between the backhaul manager and the
Agent, we can't use the fronthaul argument we used to have to fill the
radio_mac on the beerocks message header.
Since the Controller expects to receive these messages from the radio
holding the backhaul link, change the interface to the selected
interface name on the database.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>