boardfarm: added roll_logs function to prplmesh_prplwrt
Even thought the prplmesh_utils call is currently
the same for prplwrt and prplmesh_compose devices,
it might not be the case in the future, so it's
best to keep the function call abstract by placing
it inside the device class. The script location
however, it's already different, so it makes even
more sense to leave it class dependant.

Signed-off-by: David <davidmgafonso@gmail.com>