boardfarm: added roll_logs call to test teardown
By adding a roll log on each test teardown we
stop running into the problem of a log line not
being found because the log was just rolled for
being too big.
It also becomes easier to locate the log for
each test, not only by relying on the timestamp
but through its name as a subfix.

Signed-off-by: David <davidmgafonso@gmail.com>