agent: Create a single Agent socket on the Backhaul manager
Until now, the backhaul manager held a son_slave socket for each radio.
Since now there is only a single agent instance, no reason to do that
anymore.
The socket was on a container called "slaves_sockets" which help an
additional radio information, such as interfaces names.
The socket was removed from this container and moved to be a separated
backhaul manager class member. Since the container does not hold a
socket anymore, it has been renamed to "m_radios_info".

Change everywhere in the backhaul manager and the agent tasks to use
the new agent socket and radios_info container with the new name.

Additional changes:

1. We used to save a slave socket as a class member called
"m_unassociated_measurement_slave_soc" which belonged to backhaul
manager measurement flow. We used it to not bother all slaves instances
with the measurement flow. Since we have only a single agent instance,
it is no longer required and has been removed.

2. Removed conversion functions to/from slave socket which are no longer
needed since we have a single Agent socket.

3. Removed "m_disabled_slave_sockets" class member which used to hold
the ZWDFS slave socket. It is no longer required. Since we have a
a single Agent socket anyway, and to distinguish the radio we want to
use, a radio MAC has been set on the beerocks message header in a
previous commit.

4. On the Agent registration message, we are now only saving the agent
socket. All message members which used in the past to add the slave
information on the "slave sockets" are no longer used in the message
handler. Instead, we will now load all the radio information from the
AgentDB when receiving the BACKHAUL_ENABLE message.

5. The function "finalize_slaves_connect_state()" is no longer able to
receive a specific socket to send the "backhaul connected" notification.
Its implementation has been simplified since now we only have one socket
to send the notification to.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>