agent: Remove backhaul manager "Pending Enable" mechanism
Until now the Backhaul manager was able to be enabled separately from
each radio slave. Since now we have only Agent, we can't do that anymore
and the ENABLE message is now ENABLE for all radios.
Remove the "pending enable" flow and its associated class members.

In addition:
1. The endable flow was coupled with the "slave finalization" flow
which allowed to send a connected notification to each radio agent
separately. Remove it too as part of this commit.
Note that after removing it, we don't need the state "PRE_OPERATIONAL"
anymore so removed it too.

2. As part of the backhaul enable flow, the backhaul was able to send
to the radio agent that it is "busy" when not all the radio agents sent
the enable message yet, so the son_slave will be able to initiate the
enable flow again, if to other agents it takes more time do it for some
reason. Remove this message, since it is not required anymore.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>