agent: Change Agent tear down request from backhaul manager
Until now the Backhaul manager sent a teardown request (if required by
the UCC) to each radio agent separately, the radio agent forwarded
it to the correct AP manager.
Now we have a single Agent socket so change it so the backhaul manager
send it once to the Agent socket, and the Agent forward it to each AP
manager.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>