agent: Notify AP DISABLE to backhaul manager
The channel selection needs to know if AP was disabled.
Until now the event has been sent to the channel selection task when
the radio agent has disconnected. If a radio received AP DISABLED,
it disconnected the socket to the backhaul manager as part of the slave
reset, and the backhaul manager created an AP DISABLED event to the
channel selection task.
Since now it is not possible to know on the backhaul manager
disconnection handler if a specific radio got AP DISABLED, create a new
message that says explicitly if such event has occurred and for each
radio.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>