agent: Postponed attach timeout if on DFS state
dwpal: Postponed attach timeout if on DFS state

Apparently, after the changes in the previous commits, the flow of the
code is faster. It is observed that the Agent thread gets the UCC
"dev_reset_default" faster than before which is before the AP finishes
the attachment FSM.
The attachment fails on timeout since the radio is in ACS mode, and it
takes more than 15 seconds (current defined timeout) for it to attach.
When it fails, the thread is closed and the Agent get disconnect
event and go to the "STOPPED" state since the "stop_on_failure"
counter is incremented.
Before, it didn't happen since the Agent got "Backhaul Disconnected"
event triggered by "dev_reset_default", before the AP failed on the
attach process.

Change the dwpal code to not fail BWL attachment if we are waiting for
the initial ACS to finish.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>