db/client_association_task: save sta capability info in db
When with a first station association for the first time received
the Client Capability Report message it's desired to save this information
to the controller database. So subsequent associations can fetch the capabilities
from the database instead of re-querying the Client Capability Report messages.
This information is saved in the controller db and not in the Station
because Station is not created in that moment yet.

PPM-792
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>