slave: helpers: get ethernet interface mediatype from max advertised speed
Pblm: using the current link speed of ethernet interfaces, in order to classify them
(fastEthernet, gigabitEthernet).
But, the link speed is unknown when the port is not connected to a station.
Besides, it depends on the auto-negotiation with the peer station, which
leads to variable value that does not indicate the real port mediatype.

Fix: deduce the mediatype from the maximum advertised link mode speed,
which is a permanent value and does not depend on negotiation with peer
station

https://jira.prplfoundation.org/browse/PPM-1672

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>