db: remove obsolete way of adding association event datao bject
With new implementation of adding AssociationEventData objects,
the previous one can be removed as well as m_assoc_inx storage.
With help of which last added AssociationEventData was accessed.

Capabilities testing was removed from nbapi_associa‎tion_event.py
Due to a problem with Association frame.
Need to return it in PPM-1773.

PPM-1272
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>