broker_client: on socket disconnect return false to stop the process
Preparative commit.
It is seen that when transport process is killed, the controller and
agent processes are aborted.
We wish instead to end the processes gracefull in such a case.

To support this we need the broker-client to return false from the
disconnect handler so the event-loop will stop and controller process
will stop as well gracefully - since the controller cannot recover the
broker socket.

PPM-1675.

Signed-off-by: Adam Dov <adov@maxlinear.com>