tools; builder: openwrt: add --shell
When debugging failing builds, it can be useful to drop into a shell
into the build container, in order to debug it interactively.

Running the container manually is not convenient, as one would have to
repeat all the options that are given when the script does the build.

Add a `--shell` option, which stats a shell inside the container
instead of doing the build. From there, `./build_scripts/build.sh` can
be invoked interactively.

For it to work, the `-t` option has to be passed to docker as well (we
can't use it unconditionally as the option can't be used when doing a
non-interactive CI build).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>