tlvf: Fix tlvf conversion function
On many TLVs we define string like this:
  some_str:
    _type: char
    _length: [ CONST_LENGTH ]

If we set it with a string smaller than CONST_LENGTH, the string in the
buffer will have a null terminator from the string end to the
CONST_LENGTH.

The auto-generated functions that return string object build it by
passing the size of it as a second argument.
This cause to the string having a bigger length than it actually has,
containing all the null terminators.
This is dangerous since the string object returned by the function will
not be equal to a string that has the correct length.
If using the incorrect string with containers like a map for instance,
it could create a new element in the container although one is already
exist.
Fix it by eliminating the redundant null terminators from the string
object.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>