cmake: bump minimum version to 3.12
cmake versions below 3.12 have no way to reliably find a python3
interpreter when a python version without suffix is available in the
environment.

See the note [1]:
A call to find_package(PythonInterp ${V}) for python version V may
find a python executable with no version suffix. In this case no
attempt is made to avoid python executables from other versions. Use
FindPython3, FindPython2 or FindPython instead.

See also the discussion in [2].

FindPython3 has been introduced in cmake to solve this issue, but it's
only available in cmake 3.12+.

Since we need it to reliably find python3, bump the minimum version we
require to 3.12.

[1]: https://cmake.org/cmake/help/latest/module/FindPythonInterp.html
[2]: https://reviews.llvm.org/D64881

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>